Migrate Event route to Event module
